+++ unlock-js/src/__tests__/v0/partialWithdrawFromLock.test.js
@@ -1,5 +1,5 @@
 import * as UnlockV0 from 'unlock-abi-0'
-import * as utils from '../../utils.ethers'
+import * as utils from '../../utils'
 import Errors from '../../errors'
 import TransactionTypes from '../../transactionTypes'
 import NockHelper from '../helpers/nockHelper'
