+++ unlock-js/src/__tests__/v01/partialWithdrawFromLock.test.js
@@ -1,5 +1,5 @@
 import * as UnlockV01 from 'unlock-abi-0-1'
-import * as utils from '../../utils.ethers'
+import * as utils from '../../utils'
 import Errors from '../../errors'
 import TransactionTypes from '../../transactionTypes'
 import NockHelper from '../helpers/nockHelper'
