+++ unlock-js/src/__tests__/v02/createLock.test.js
@@ -1,6 +1,6 @@
 import { ethers } from 'ethers'
 import * as UnlockV02 from 'unlock-abi-0-2'
-import * as utils from '../../utils.ethers'
+import * as utils from '../../utils'
 import Errors from '../../errors'
 import TransactionTypes from '../../transactionTypes'
 import NockHelper from '../helpers/nockHelper'
